fix: 방 상세 수정 todo와 room에서 사용할 UserInfoRoomDto 생성 방 상세 dto 수정

+++ src/main/java/com/example/groom/domain/room/dto/RoomDto.java
@@ -14,7 +14,7 @@ public class RoomDto {
     private String description;
     private String mainImageUrl;
     private Long maxPeopleNumber;
-    private Long nowPeopleNumber;
+    private int nowPeopleNumber;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
